requirements.txt
Django
psycopg2-binary
django-minio-storage
Pillow
djangorestframework
PyJwt
django-cors-headers
channels
channels-redis
asgi_redis
channels_redis
